Redesign channel model execution
The previous execution model performance heavily
relied upon the sampling rate and the time resolution of
the simulation. In some cases the simulation was
not advancing at all.

This commit uses a different approach that allow
the simulation to advance without been affected
by the sampling rate and/or the time resolution

Signed-off-by: Manolis Surligas <surligas@gmail.com>